+++ x/zkauth/types/errors.go
@@ -4,6 +4,9 @@ import (
 	sdkerrors "github.com/Finschia/finschia-sdk/types/errors"
 )
 
+const zkAuthCodespace = ModuleName
+
 var (
-	ErrSample = sdkerrors.Register(ModuleName, 1100, "sample error")
+	ErrInvalidZkAuthInputs = sdkerrors.Register(zkAuthCodespace, 2, "invalid zkauth inputs")
+	ErrSample              = sdkerrors.Register(zkAuthCodespace, 1100, "sample error")
 )